fix bug

+++ include/mcl/stdtype.h
@@ -7,8 +7,8 @@
 #include <stdbool.h>
 #include <string.h>
 
-static const uint32_t MCL_UINT8_MAX  = 0xFF;
-static const uint32_t MCL_UINT16_MAX = 0xFFFF;
+static const uint8_t  MCL_UINT8_MAX  = 0xFF;
+static const uint16_t MCL_UINT16_MAX = 0xFFFF;
 static const uint32_t MCL_UINT32_MAX = 0xFFFFFFFF;
 static const uint64_t MCL_UINT64_MAX = 0xFFFFFFFFFFFFFFFF;
